refactor: migrate authentication mutations to a dedicated Auth data provider and update related tests
