+++ client/src/data-provider/mutations.ts
@@ -1,20 +1,16 @@
 import {
   Constants,
-  LocalStorageKeys,
   InfiniteCollections,
   defaultAssistantsVersion,
   ConversationListResponse,
 } from 'librechat-data-provider';
-import { useResetRecoilState } from 'recoil';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { dataService, MutationKeys, QueryKeys, defaultOrderQuery } from 'librechat-data-provider';
 import type * as t from 'librechat-data-provider';
 import type { InfiniteData, UseMutationResult } from '@tanstack/react-query';
 import useUpdateTagsInConvo from '~/hooks/Conversations/useUpdateTagsInConvo';
 import { updateConversationTag } from '~/utils/conversationTags';
-import useClearStates from '~/hooks/Config/useClearStates';
 import { normalizeData } from '~/utils/collection';
-import store from '~/store';
 import {
   useConversationTagsQuery,
   useConversationsInfiniteQuery,
@@ -693,28 +689,6 @@ export const useDeletePresetMutation = (
   });
 };
 
-/* login/logout */
-export const useLogoutUserMutation = (
-  options?: t.LogoutOptions,
-): UseMutationResult<unknown, unknown, undefined, unknown> => {
-  const queryClient = useQueryClient();
-  const clearStates = useClearStates();
-  const resetDefaultPreset = useResetRecoilState(store.defaultPreset);
-
-  return useMutation([MutationKeys.logoutUser], {
-    mutationFn: () => dataService.logout(),
-    ...(options || {}),
-    onSuccess: (...args) => {
-      resetDefaultPreset();
-      clearStates();
-      queryClient.removeQueries();
-      localStorage.removeItem(LocalStorageKeys.LAST_MODEL);
-      localStorage.removeItem(LocalStorageKeys.FILES_TO_DELETE);
-      options?.onSuccess?.(...args);
-    },
-  });
-};
-
 /* Avatar upload */
 export const useUploadAvatarMutation = (
   options?: t.UploadAvatarOptions,
@@ -727,27 +701,6 @@ export const useUploadAvatarMutation = (
   return useMutation([MutationKeys.avatarUpload], {
     mutationFn: (variables: FormData) => dataService.uploadAvatar(variables),
     ...(options || {}),
-  });
-};
-
-export const useDeleteUserMutation = (
-  options?: t.MutationOptions<unknown, undefined>,
-): UseMutationResult<unknown, unknown, undefined, unknown> => {
-  const queryClient = useQueryClient();
-  const clearStates = useClearStates();
-  const resetDefaultPreset = useResetRecoilState(store.defaultPreset);
-
-  return useMutation([MutationKeys.deleteUser], {
-    mutationFn: () => dataService.deleteUser(),
-    ...(options || {}),
-    onSuccess: (...args) => {
-      resetDefaultPreset();
-      clearStates();
-      queryClient.removeQueries();
-      localStorage.removeItem(LocalStorageKeys.LAST_MODEL);
-      localStorage.removeItem(LocalStorageKeys.FILES_TO_DELETE);
-      options?.onSuccess?.(...args);
-    },
   });
 };
